+++ src/renderer/hooks/use-check-for-updates.ts
@@ -2,12 +2,15 @@ import { useQuery } from '@tanstack/react-query';
 import isElectron from 'is-electron';
 import { useEffect, useState } from 'react';
 
+import { useWindowSettings } from '/@/renderer/store';
+
 const CHECK_FOR_UPDATES_INTERVAL_MS = 6 * 60 * 60 * 1000;
 
 const utils = isElectron() ? window.api?.utils : null;
 
 export const useCheckForUpdates = () => {
     const [enablePeriodicCheck, setEnablePeriodicCheck] = useState(false);
+    const { disableAutoUpdate } = useWindowSettings();
 
     // We want to skip the first check since it's already checked in the main process when the app is started
     useEffect(() => {
@@ -17,6 +20,7 @@ export const useCheckForUpdates = () => {
 
     const isEnabled =
         enablePeriodicCheck &&
+        !disableAutoUpdate &&
         Boolean(isElectron() && utils?.checkForUpdates && !utils?.disableAutoUpdates?.());
 
     return useQuery({