+++ src/renderer/hooks/use-native-menu-sync.tsx
@@ -5,9 +5,9 @@ import { useTranslation } from 'react-i18next';
 
 import packageJson from '../../../package.json';
 
+import { openAboutModal } from '/@/renderer/about-modal';
 import { ServerList } from '/@/renderer/features/servers/components/server-list';
 import { openSettingsModal } from '/@/renderer/features/settings/utils/open-settings-modal';
-import { openAboutModal } from '/@/renderer/about-modal';
 import { openReleaseNotesModal } from '/@/renderer/release-notes-modal';
 import {
     useAppStore,